uploaded category.php (news page) & minor changes of header.php and footer.php (read description)
-better indentation, more human readable

-added my functions for the category.php into the functions.php (mainly
for the news page)

-category.php is basically responsive, but need to work with the
designer to make it look better

-strengthened the html structure (originally missing the opening body
tags etc)

-better practice of wordpress, using <?php bloginfo('name'); ?> to get
the site title instead of hardcoding it etc

PS: Guys, I think we did not utilize github well enough. What we suppose
to do is to work on it locally and push changes to github. I believe
this is the most efficient way to do it. Please stop modifying files
directly on the server. It is so hard for others to keep track on the
progress. Whenever we complete something, we should push it to github
and let others know what we have completed. Personally I failed to do
it. However, I think this is the best way to work in collaborative
working environment. And we could definitely make a good website
together. Thank you.

These changes didn't change the layout of the website except for the
footer menu, I don't know what goes wrong so far.